PUT new file on WebDAV server

PUTs new file into a WebDAV drive, in the HTTP sense: Existing files at the same url are overwritten. Will retry as per httr::RETRY().

webdav_put(local, remote, user, password, keyring)

Arguments

local
character[1], giving the path to the local file to upload.

remote
character[1], giving the path to the file to create or overwrite on the WebDAV server. Remember to include the target filename.

user
[character(1)] giving the username to authenticate with, for example "jane". Also used to retrieve from and store the password in keyring.

password
[character(1)] giving the password. Use this to substitute in an environment variable such as Sys.getenv("SECRET"). Recommended only when the environment variable is declared in an encrypted form, and when it is redacted from the logs. Do not expose your passwords in scripts or the R console. Defaults to NULL, in which case the keyring is used.

keyring
For systems that support multiple keyrings, specify the name of the keyring to use here. If NULL, then the default keyring is used. See also has_keyring_support().

Value

httr::response

Details

The request is authenticated using passhelpr::authenticate2(), with the service as the domain root from remote.